Added Combine Script as well as other assets
The other assets are used on the desk to make the area seem more spacious as well as to use the combine script to successfully combine two items. The combine script will combine the object with the script attached with another item to give us a "Combined Item" that the player will use to either put in a specific location or to make a new item needed for creation later.