Added cardboard box with items, new models, and attached scripts for drop zones to work
The cardboard box works similarly to that of level 1 in which it shoots items out for the user to grab and place around the room. Added a new pen and new pen holder prefab with the new pen holder having two pens for difference/progression within the story. Finally, ensured all the drop zones had the proper scripts attached to each of them similarly to how they were set up in level 1. Will begin working on the music player for the user to use now.