Added music player within different scene
Ensure that the OVRCameriaRig has similar properties to that of the OVRCameraRigInteration so that it can use the ray's to click on the UI within the music player. To add a song you will need to make a Song Scriptable Object (similar to how you create objects in unity), and add your audio clip and image for the song. It will take care of the time for you as well as going to next song when the song is complete. You can also rewind the song if you rewind after 3 seconds and play the song before if you rewind before 3 seconds (if there was one previous to the song). You can also skip to the next song and if you skip past the last song the list starts over from the first song. If the song naturally runs out it will go to the next song by default.