Added Music Player into the scene and teleport
Each level now has a music player with the two lofi songs attached to them to play music while the user is putting stuff away within the level. Added the level manager to the second scene so that when the player wins on last level it will say "end of demo" in the console as a log statement (should add end in final release). The 2nd level also has a teleport navmesh so that users can teleport around the room and explore as well as not be able to get on top of desks or beds with the nav mesh obstacle (added one in level 1 for the stand). Have not tested due to headset being dead, will make changes if an issue occurs from partners or when headset is charged and active.