Made fixes to drop zone checker and fade canvas
Added a script called itemChecker that will check to see if the item is grabbed or not, if it is grabbed then it will not count the area as being filled/done since it is still being held. Once let go it will then allow it to go through the rest of the OnTriggerStay function to determine the correctness of the item placement.